Split the giant attendee-merge test into a mirrored folder suite
`deno task unit-tests-report` flagged `src/shared/merge/attendee-merge.ts`
as untested: its unit test lived at the non-mirror path
`test/lib/attendee-merge.test.ts` and, at 2083 lines, was grandfathered
past the 1000-line Biome ceiling.

This moves it to the mirror path as a folder suite and, in doing so,
breaks the one giant file into focused, well-named files — each well
under 400 lines:

  test/shared/merge/attendee-merge/
    helpers.ts        — shared setup + merge helpers (not a test file)
    labels.test.ts    — the pure label/key/conflict helpers
    diff.test.ts      — buildAttendeeMergeDiff
    validate.test.ts  — validateAttendeeMergeDecision
    apply.test.ts     — applyAttendeeMerge
    repoint.test.ts   — ledger repointing on merge

Splitting a monolith surfaces duplication jscpd could not scan inside
the huge original, so the bulk of this change is extracting that
duplication into helpers rather than moving lines: a `runMerge`
build-and-apply helper with standard "Bob → Alice" PII, `buildMergeDiff`,
`postPaidSale`, `saveChoice`/`saveTextAnswer`, `createMergePair`, and
declarative `row`/`bookingItem`/`diffWith` builders for the validation
tests. All 40 tests keep their exact assertions; the folder has zero
duplication.

Two of those helpers were already duplicated elsewhere in the suite, so
they move to `#test-utils` and both call sites now share them:
`bookTestAttendee` (also used by `test/lib/questions/helpers.ts`) and
`runAndCountRoundTrips` (also used by the ticket-payment reservation
test). No source code changed.

The report now credits `attendee-merge.ts` as mirror-tested, and its
Biome max-lines exemption is dropped.
