+++ test/shared/merge/attendee-merge/apply.test.ts
@@ -1,0 +1,390 @@
+import { expect } from "@std/expect";
+import { describe, it as test } from "@std/testing/bdd";
+import { getDb, queryAll } from "#shared/db/client.ts";
+import {
+  getAttendeeAnswersByQuestion,
+  getAttendeeTextAnswers,
+} from "#shared/db/questions.ts";
+import { bookingKey } from "#shared/merge/attendee-merge.ts";
+import {
+  createTestListing,
+  describeWithEnv,
+  runAndCountRoundTrips,
+} from "#test-utils";
+import { getTestPrivateKey } from "#test-utils/crypto.ts";
+import {
+  applyMerge,
+  bookingChoice,
+  buildMergeDiff,
+  createAttendee,
+  createAttendeeOn,
+  createFreeTextQuestion,
+  createMergePair,
+  createQuestionWithAnswers,
+  getBookings,
+  oneQuestion,
+  pii,
+  postPaidSale,
+  runMerge,
+  saveChoice,
+  saveTextAnswer,
+} from "./helpers.ts";
+
+type MergePair = Awaited<ReturnType<typeof createMergePair>>;
+
+/** Merge `source` into `target` and assert the target ends up with `expected`
+ *  as its free-text answer for `questionId`. */
+const expectMergedTextAnswer = async (
+  source: MergePair["source"],
+  target: MergePair["target"],
+  questionId: number,
+  expected: string,
+) => {
+  const { result } = await runMerge({ source, target });
+  expect(result.success).toBe(true);
+  const textAnswers = await getAttendeeTextAnswers(
+    target.id,
+    await getTestPrivateKey(),
+  );
+  expect(textAnswers.get(questionId)).toBe(expected);
+};
+
+describeWithEnv("attendee merge service", { db: true }, () => {
+  describe("applyAttendeeMerge", () => {
+    test("clears check-in when copying a no-quantity source line", async () => {
+      const listing1 = await createTestListing({
+        maxAttendees: 10,
+        name: "M1",
+      });
+      const listing2 = await createTestListing({
+        maxAttendees: 10,
+        name: "M2",
+      });
+      const target = await createAttendee(listing1.id, "Alice", "a@test.com");
+      const source = await createAttendee(listing2.id, "Bob", "b@test.com");
+      // Make source's line a checked-in quantity-0 sentinel (price 0).
+      await getDb().execute({
+        args: [source.id],
+        sql: "UPDATE listing_attendees SET quantity = 0, checked_in = 1 WHERE attendee_id = ?",
+      });
+
+      const { result } = await runMerge({
+        source,
+        sourcePii: pii("Bob", "b@test.com"),
+        target,
+        targetPii: pii("Alice", "a@test.com"),
+      });
+      expect(result.success).toBe(true);
+
+      const moved = (await getBookings(target.id)).find(
+        (b) => b.listing_id === listing2.id,
+      )!;
+      expect(moved.quantity).toBe(0);
+      // The ghost line arrives with its check-in cleared.
+      expect(moved.checked_in).toBe(0);
+    });
+
+    test("applies PII and answer decisions correctly", async () => {
+      const listing1 = await createTestListing({
+        maxAttendees: 10,
+        name: "E1",
+      });
+      const listing2 = await createTestListing({
+        maxAttendees: 10,
+        name: "E2",
+      });
+
+      const { question, answers } = await createQuestionWithAnswers(
+        listing1.id,
+        "Colour?",
+        ["Red", "Blue"],
+      );
+
+      const target = await createAttendee(
+        listing1.id,
+        "Alice",
+        "alice@test.com",
+      );
+      const source = await createAttendee(listing2.id, "Bob", "bob@test.com");
+
+      await saveChoice(target.id, answers[0]!.id); // Red
+      await saveChoice(source.id, answers[1]!.id); // Blue
+
+      const { result } = await runMerge({
+        decide: () => ({
+          answers: { [String(question.id)]: "source" },
+          bookings: {},
+          money: {},
+          pii: { email: "target", name: "source" },
+        }),
+        questions: oneQuestion(question, answers),
+        source,
+        target,
+      });
+
+      expect(result.success).toBe(true);
+      expect(result.summary.piiFieldsFromSource).toEqual(["name"]);
+      expect(result.summary.answersTakenFromSource).toBe(1);
+      expect(result.summary.bookingsMoved).toBe(1); // listing2 moved to target
+
+      // Verify answers were updated
+      const finalAnswers = await getAttendeeAnswersByQuestion(target.id);
+      expect(finalAnswers.get(question.id)?.answerId).toBe(answers[1]!.id);
+
+      // Verify source deleted
+      const sourceRows = await queryAll<{ id: number }>(
+        "SELECT id FROM attendees WHERE id = ?",
+        [source.id],
+      );
+      expect(sourceRows.length).toBe(0);
+
+      // Verify target has both listing links
+      const listingLinks = await queryAll<{ listing_id: number }>(
+        "SELECT listing_id FROM listing_attendees WHERE attendee_id = ?",
+        [target.id],
+      );
+      expect(listingLinks.map((r) => r.listing_id).sort()).toEqual(
+        [listing1.id, listing2.id].sort(),
+      );
+    });
+
+    test("merges many paid duplicate bookings in a bounded number of round-trips", async () => {
+      // Regression: each discarded paid booking posts a decision-17 reversal leg.
+      // Posting them one-leg-at-a-time inside an interactive transaction held the
+      // write lock open for several round-trips PER leg, so a big merge could blow
+      // the primary's transaction timeout. The whole merge must commit as one batch
+      // — O(1) round-trips regardless of how many bookings the person has.
+      const N = 12;
+      // Sequential: each createTestListing runs an authenticated request that
+      // mints a session, so building them concurrently would collide session
+      // tokens — the round-trip count we assert on is the merge, not the setup.
+      const listings: Awaited<ReturnType<typeof createTestListing>>[] = [];
+      for (let i = 0; i < N; i++) {
+        listings.push(await createTestListing({ maxAttendees: 10 }));
+      }
+      const listingIds = listings.map((l) => l.id);
+      const target = await createAttendeeOn(
+        listingIds,
+        "Alice",
+        "alice@test.com",
+      );
+      const source = await createAttendeeOn(listingIds, "Bob", "bob@test.com");
+
+      // Give every source booking a paid sale leg, so each same-listing duplicate
+      // is a paid conflict that needs a money decision (a reversal leg on merge).
+      for (let i = 0; i < N; i++) {
+        await postPaidSale({
+          attendeeId: source.id,
+          eventGroup: `evt${i}`,
+          listingId: listings[i]!.id,
+        });
+        // Link the booking row to its ledger event so the merge sees it as paid.
+        await getDb().execute({
+          args: [`evt${i}`, source.id, listings[i]!.id],
+          sql: "UPDATE listing_attendees SET ledger_event_group = ? WHERE attendee_id = ? AND listing_id = ?",
+        });
+      }
+
+      const sourcePii = pii("Bob", "bob@test.com");
+      const targetPii = pii("Alice", "alice@test.com");
+      const diff = await buildMergeDiff({
+        source,
+        sourcePii,
+        target,
+        targetPii,
+      });
+
+      // Every listing is a duplicate conflict: keep the target row and write off
+      // the discarded source booking's cash.
+      const bookings: Record<string, "keep_target"> = {};
+      const money: Record<string, "writeoff"> = {};
+      for (const l of listings) {
+        bookings[`${l.id}:null`] = "keep_target";
+        money[`${l.id}:null`] = "writeoff";
+      }
+
+      const decision = {
+        answers: {},
+        bookings,
+        money,
+        pii: {},
+        version: diff.version,
+      };
+      // Count how many DB round-trips the apply itself takes, ignoring setup.
+      const { value: result, roundTrips } = await runAndCountRoundTrips(() =>
+        applyMerge({ decision, diff, source, sourcePii, target, targetPii }),
+      );
+
+      expect(result.success).toBe(true);
+      expect(result.summary.bookingsWrittenOff).toBe(N);
+      // The N reversal legs land in one batch, so the merge's round-trips don't
+      // scale with N (an interactive per-leg post would be ~3N and trip the guard).
+      expect(roundTrips).toBeLessThanOrEqual(10);
+    });
+
+    test("preserves the target's free-text answers through a merge", async () => {
+      // Regression: the merge re-saves only the target's choice answers, which
+      // deletes every attendee_answers row for the target. Without carrying the
+      // free-text answers through, those text rows were silently wiped.
+      const { target, source } = await createMergePair();
+      const textQuestion = await createFreeTextQuestion();
+      // The target owns the free-text answer that must survive the merge.
+      await saveTextAnswer(target.id, textQuestion.id, "Coeliac");
+
+      await expectMergedTextAnswer(source, target, textQuestion.id, "Coeliac");
+    });
+
+    test("adopts a source-only free-text answer in a merge", async () => {
+      // Source-only choice answers are adopted automatically; a source-only
+      // text answer must be too, rather than vanishing when the source is
+      // deleted.
+      const { target, source } = await createMergePair();
+      const textQuestion = await createFreeTextQuestion();
+      // Only the source owns the answer; the merge must adopt it, not drop it.
+      await saveTextAnswer(source.id, textQuestion.id, "Vegan");
+
+      await expectMergedTextAnswer(source, target, textQuestion.id, "Vegan");
+    });
+
+    test("clears answers when decision is clear", async () => {
+      const { listing, target, source } = await createMergePair();
+      const { question, answers } = await createQuestionWithAnswers(
+        listing.id,
+        "Size?",
+        ["S", "L"],
+      );
+
+      await saveChoice(target.id, answers[0]!.id);
+      await saveChoice(source.id, answers[1]!.id);
+
+      const { result } = await runMerge({
+        decide: () => ({
+          answers: { [String(question.id)]: "clear" },
+          bookings: {},
+          money: {},
+          pii: {},
+        }),
+        questions: oneQuestion(question, answers),
+        source,
+        target,
+      });
+
+      expect(result.summary.answersCleared).toBe(1);
+      const finalAnswers = await getAttendeeAnswersByQuestion(target.id);
+      expect(finalAnswers.has(question.id)).toBe(false);
+    });
+
+    test("adopts source answers when target has none", async () => {
+      const { listing, target, source } = await createMergePair();
+      const { question, answers } = await createQuestionWithAnswers(
+        listing.id,
+        "Meal?",
+        ["Chicken", "Fish"],
+      );
+
+      // Only source has answer
+      await saveChoice(source.id, answers[1]!.id);
+
+      const { result } = await runMerge({
+        questions: oneQuestion(question, answers),
+        source,
+        target,
+      });
+
+      expect(result.summary.answersTakenFromSource).toBe(1);
+      const finalAnswers = await getAttendeeAnswersByQuestion(target.id);
+      expect(finalAnswers.get(question.id)?.answerId).toBe(answers[1]!.id);
+    });
+
+    test("handles duplicate booking with keep_target decision", async () => {
+      const { listing, target, source } = await createMergePair({
+        sameListing: true,
+      });
+
+      const key = bookingKey(listing.id, null, 0, 0);
+      const { diff, result } = await runMerge({
+        decide: bookingChoice(key, "keep_target"),
+        source,
+        target,
+      });
+
+      expect(diff.bookingItems[0]!.conflictClass).toBe("duplicate");
+      expect(result.summary.bookingsSkipped).toBe(1);
+      expect(result.summary.bookingsMoved).toBe(0);
+
+      // Target still has exactly 1 booking
+      const links = await queryAll<{ listing_id: number }>(
+        "SELECT listing_id FROM listing_attendees WHERE attendee_id = ?",
+        [target.id],
+      );
+      expect(links.length).toBe(1);
+    });
+
+    test("replaces target booking with take_source decision", async () => {
+      const { listing, target, source } = await createMergePair({
+        sameListing: true,
+      });
+
+      // Update source booking to have different quantity to create conflicting_metadata
+      await queryAll(
+        "UPDATE listing_attendees SET quantity = 5 WHERE attendee_id = ?",
+        [source.id],
+      );
+
+      const key = bookingKey(listing.id, null, 0, 0);
+      const { diff, result } = await runMerge({
+        decide: bookingChoice(key, "take_source"),
+        source,
+        target,
+      });
+
+      expect(diff.bookingItems[0]!.conflictClass).toBe("conflicting_metadata");
+      expect(result.summary.bookingsReplacedTarget).toBe(1);
+
+      // Target's booking should now have qty 5
+      const links = await queryAll<{ quantity: number }>(
+        `SELECT quantity
+         FROM listing_attendees
+         WHERE attendee_id = ?
+           AND listing_id = ?`,
+        [target.id, listing.id],
+      );
+      expect(links.length).toBe(1);
+      expect(links[0]!.quantity).toBe(5);
+    });
+
+    test("returns accurate summary counts", async () => {
+      const listing1 = await createTestListing({
+        maxAttendees: 10,
+        name: "E1",
+      });
+      const listing2 = await createTestListing({
+        maxAttendees: 10,
+        name: "E2",
+      });
+
+      const target = await createAttendee(listing1.id, "Alice");
+      const source = await createAttendee(listing2.id, "Bob");
+
+      const { diff, result } = await runMerge({
+        decide: () => ({
+          answers: {},
+          bookings: {},
+          money: {},
+          pii: { name: "source" },
+        }),
+        source,
+        target,
+      });
+
+      // Source is on listing2 only, target on listing1 — no conflicts, 1 moveable
+      expect(diff.bookingItems.length).toBe(1);
+      expect(diff.bookingItems[0]!.conflictClass).toBe("moveable");
+      expect(result.success).toBe(true);
+      expect(result.summary.piiFieldsFromSource).toEqual(["name"]);
+      expect(result.summary.bookingsMoved).toBe(1);
+      expect(result.summary.bookingsSkipped).toBe(0);
+      expect(result.summary.bookingsReplacedTarget).toBe(0);
+    });
+  });
+});
