+++ test/test-utils.ts
@@ -24,6 +24,7 @@ export * from "./test-utils/modifiers.ts";
 export * from "./test-utils/order-journey.ts";
 export * from "./test-utils/packages.ts";
 export * from "./test-utils/parents.ts";
+export * from "./test-utils/query-log.ts";
 export * from "./test-utils/random.ts";
 export * from "./test-utils/servicing.ts";
 export * from "./test-utils/session.ts";
